Resolve #3 - Implement Region services and partially tests for it

+++ app/data/entities/LocationEntity.scala
@@ -1,6 +1,6 @@
 package data.entities
 
 case class LocationEntity(
-                           latitude: Float,
-                           longitude: Float
+                           latitude: Double,
+                           longitude: Double
                          )
