+++ app/data/entities/ZoneEntity.scala
@@ -2,5 +2,5 @@ package data.entities
 
 case class ZoneEntity(
                        center: LocationEntity,
-                       radius: Float
+                       radius: Double
                      )
